+++ lib/groups/index.ts
@@ -1,3 +1,4 @@
 export * from './types';
 export * from './permissions';
 export * from './schema';
+export * from './assignments';